netatmo/api/requirements.txt
lnetatmo
psycopg2-binary